Change behavior of refresh roles
