+++ main.py
@@ -5,7 +5,7 @@ from dotenv import load_dotenv
 import disnake
 from disnake.ext import commands, tasks
 from src import keep_alive
-from src.utils.Codeforces import guild_funcs
+from src.utils import guild_funcs
 
 load_dotenv()
 guilds = [int(v) for v in environ.get("TEST_GUILDS").split(",")]
